Refactored policy parsing to separate each SC option handling, improved failure logging and handling. Added unit tests for policy parsing valid path.
