+++ .github/workflows/formatting.yaml
@@ -45,7 +45,7 @@ jobs:
       - name: Set up OS
         run: |
           sudo apt-get update && sudo apt-get install -y \
-              clang-format python3-pip
+              clang-format jq python3-pip
           pip3 install autopep8 licenseheaders
       - name: Check code formatting
         run: ./build.sh check-format
