+++ mock-bpa-test/key_set_1.json
@@ -1,19 +1,19 @@
 {
-    "keys": [
-        {
-            "kty": "oct",
-            "kid": "9100",
-            "k": "GisaKxorGisaKxorGisaKw=="
-        },
-        {
-            "kty": "oct",
-            "kid": "9102",
-            "k": "cXdlcnR5dWlvcGFzZGZnaA=="
-        },
-        {
-            "kty": "oct",
-            "kid": "9103",
-            "k": "YWJjZGVmZ2hpamtsbW5vcA=="
-        }
-    ]
+  "keys": [
+    {
+      "kty": "oct",
+      "kid": "9100",
+      "k": "GisaKxorGisaKxorGisaKw=="
+    },
+    {
+      "kty": "oct",
+      "kid": "9102",
+      "k": "cXdlcnR5dWlvcGFzZGZnaA=="
+    },
+    {
+      "kty": "oct",
+      "kid": "9103",
+      "k": "YWJjZGVmZ2hpamtsbW5vcA=="
+    }
+  ]
 }
