+++ resources/apply_format.sh
@@ -43,5 +43,19 @@ else
 fi
 clang-format --style=file -i $ARGS
 
+# Test configs
+for FILENAME in $(find . \( -path './deps' -prune -o -name '*.json' \) -a -type f)
+do
+    if jq . <"${FILENAME}" >"${FILENAME}.new"
+    then
+        mv "${FILENAME}.new" "${FILENAME}"
+    else
+        ERR=$?
+        echo "Failed to format ${FILENAME}" >/dev/stderr
+        rm "${FILENAME}.new"
+        exit $ERR
+    fi
+done
+
 # Python test fixtures
 autopep8 --max-line-length=100 -ir mock-bpa-test/
