+++ test/test_PolicyParser/empty.json
@@ -1,0 +1,3 @@
+{
+  "policyrule_set": []
+}
