+++ test/test_PolicyParser/validSC2.json
@@ -2,53 +2,6 @@
   "policyrule_set": [
     {
       "policyrule": {
-        "desc": "Integrity source rule",
-        "filter": {
-          "rule_id": "1",
-          "role": "s",
-          "tgt": 1,
-          "loc": "appin",
-          "sc_id": 1
-        },
-        "spec": {
-          "svc": "bib-integrity",
-          "sc_id": 1,
-          "sc_parms": [
-            {
-              "id": "key_name",
-              "value": "9100"
-            },
-            {
-              "id": "sha_variant",
-              "value": "7"
-            },
-            {
-              "id": "scope_flags",
-              "value": "0"
-            },
-            {
-              "id": "key_wrap",
-              "value": "0"
-            }
-          ]
-        },
-        "es_ref": "d_integrity",
-        "policy_action_on_fail": "delete_bundle"
-      },
-      "event_set": {
-        "es_ref": "d_integrity",
-        "events": [
-          {
-            "event_id": "sop_processed",
-            "actions": [
-              "remove_all_target_sops"
-            ]
-          }
-        ]
-      }
-    },
-    {
-      "policyrule": {
         "desc": "Confidentiality source rule",
         "filter": {
           "rule_id": "2",
